// src/types/core/api.ts
export type AppLevel = -1 | 0 | 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8 | 9
export type AppState = 'working' | 'inprogress' | 'thirdparty'
export type AppUpstream = {
  license?: string | null
  license_url?: string | null
  website?: string | null
  demo?: string | null
  admindoc?: string | null
  userdoc?: string | null
  code?: string | null
  cpe?: string | null
}
export type AppMinManifest = {
  packaging_format: 1 | 2
  id: string
  name: string
  description: Translation
  version: string
  maintainers: string[]
  integration: {
    architectures: string | string[]
    ldap: 'not_relevant' | boolean
    sso: 'not_relevant' | boolean
    multi_instance: boolean
    disk: string
    ram: { build: string; runtime: string }
    yunohost: string
  }
  antifeatures?: Obj<Translation>
  upstream: AppUpstream
}

export type AppManifest = AppMinManifest & {
  upstream: AppUpstream
  install: AnyOption[]
  doc: {
    DESCRIPTION?: Translation
    ADMIN?: Translation
  } & Obj<Translation>
  notifications: {
    PRE_INSTALL: Obj<Translation> | null
    POST_INSTALL: Obj<Translation> | null
    PRE_UPGRADE: Obj<Translation> | null
    POST_UPGRADE: Obj<Translation> | null
  }
  requirements: Record<
    'required_yunohost_version' | 'arch' | 'install' | 'disk' | 'ram',
    {
      passed: boolean
      error: string
    }
  >
  resources: Obj
  remote: { type: string; url: string; branch: string; revision: string }
  lastUpdate: number
  quality: { level: AppLevel; state: AppState }
  antifeatures: string[]
  potential_alternative_to: string[]
  screenshot: string | null
}